+++ extern/diskann/DiskANN/apps/test_insert_deletes_consolidate.cpp
@@ -27,22 +27,22 @@ namespace po = boost::program_options;
 // load_aligned_bin modified to read pieces of the file, but using ifstream
 // instead of cached_ifstream.
 template <typename T>
-inline void load_aligned_bin_part(const std::string &bin_file, T *data, size_t offset_points, size_t points_to_read)
+inline void load_aligned_bin_part(const std::string &bin_file, T *data, uint64_t offset_points, uint64_t points_to_read)
 {
     diskann::Timer timer;
     std::ifstream reader;
     reader.exceptions(std::ios::failbit | std::ios::badbit);
     reader.open(bin_file, std::ios::binary | std::ios::ate);
-    size_t actual_file_size = reader.tellg();
+    uint64_t actual_file_size = reader.tellg();
     reader.seekg(0, std::ios::beg);
 
     int npts_i32, dim_i32;
     reader.read((char *)&npts_i32, sizeof(int));
     reader.read((char *)&dim_i32, sizeof(int));
-    size_t npts = (uint32_t)npts_i32;
-    size_t dim = (uint32_t)dim_i32;
+    uint64_t npts = (uint32_t)npts_i32;
+    uint64_t dim = (uint32_t)dim_i32;
 
-    size_t expected_actual_file_size = npts * dim * sizeof(T) + 2 * sizeof(uint32_t);
+    uint64_t expected_actual_file_size = npts * dim * sizeof(T) + 2 * sizeof(uint32_t);
     if (actual_file_size != expected_actual_file_size)
     {
         std::stringstream stream;
@@ -64,9 +64,9 @@ inline void load_aligned_bin_part(const std::string &bin_file, T *data, size_t o
 
     reader.seekg(2 * sizeof(uint32_t) + offset_points * dim * sizeof(T));
 
-    const size_t rounded_dim = ROUND_UP(dim, 8);
+    const uint64_t rounded_dim = ROUND_UP(dim, 8);
 
-    for (size_t i = 0; i < points_to_read; i++)
+    for (uint64_t i = 0; i < points_to_read; i++)
     {
         reader.read((char *)(data + i * rounded_dim), dim * sizeof(T));
         memset(data + i * rounded_dim + dim, 0, (rounded_dim - dim) * sizeof(T));
@@ -77,8 +77,8 @@ inline void load_aligned_bin_part(const std::string &bin_file, T *data, size_t o
     std::cout << "Read " << points_to_read << " points using non-cached reads in " << elapsedSeconds << std::endl;
 }
 
-std::string get_save_filename(const std::string &save_path, size_t points_to_skip, size_t points_deleted,
-                              size_t last_point_threshold)
+std::string get_save_filename(const std::string &save_path, uint64_t points_to_skip, uint64_t points_deleted,
+                              uint64_t last_point_threshold)
 {
     std::string final_path = save_path;
     if (points_to_skip > 0)
@@ -92,8 +92,8 @@ std::string get_save_filename(const std::string &save_path, size_t points_to_ski
 }
 
 template <typename T, typename TagT>
-void insert_till_next_checkpoint(diskann::AbstractIndex &index, size_t start, size_t end, int32_t thread_count, T *data,
-                                 size_t aligned_dim)
+void insert_till_next_checkpoint(diskann::AbstractIndex &index, uint64_t start, uint64_t end, int32_t thread_count, T *data,
+                                 uint64_t aligned_dim)
 {
     diskann::Timer insert_timer;
 
@@ -109,14 +109,14 @@ void insert_till_next_checkpoint(diskann::AbstractIndex &index, size_t start, si
 
 template <typename T, typename TagT>
 void delete_from_beginning(diskann::AbstractIndex &index, diskann::IndexWriteParameters &delete_params,
-                           size_t points_to_skip, size_t points_to_delete_from_beginning)
+                           uint64_t points_to_skip, uint64_t points_to_delete_from_beginning)
 {
     try
     {
         std::cout << std::endl
                   << "Lazy deleting points " << points_to_skip << " to "
                   << points_to_skip + points_to_delete_from_beginning << "... ";
-        for (size_t i = points_to_skip; i < points_to_skip + points_to_delete_from_beginning; ++i)
+        for (uint64_t i = points_to_skip; i < points_to_skip + points_to_delete_from_beginning; ++i)
             index.lazy_delete(static_cast<TagT>(i + 1)); // Since tags are data location + 1
         std::cout << "done." << std::endl;
 
@@ -137,14 +137,14 @@ void delete_from_beginning(diskann::AbstractIndex &index, diskann::IndexWritePar
 }
 
 template <typename T>
-void build_incremental_index(const std::string &data_path, diskann::IndexWriteParameters &params, size_t points_to_skip,
-                             size_t max_points_to_insert, size_t beginning_index_size, float start_point_norm,
-                             uint32_t num_start_pts, size_t points_per_checkpoint, size_t checkpoints_per_snapshot,
-                             const std::string &save_path, size_t points_to_delete_from_beginning,
-                             size_t start_deletes_after, bool concurrent)
+void build_incremental_index(const std::string &data_path, diskann::IndexWriteParameters &params, uint64_t points_to_skip,
+                             uint64_t max_points_to_insert, uint64_t beginning_index_size, float start_point_norm,
+                             uint32_t num_start_pts, uint64_t points_per_checkpoint, uint64_t checkpoints_per_snapshot,
+                             const std::string &save_path, uint64_t points_to_delete_from_beginning,
+                             uint64_t start_deletes_after, bool concurrent)
 {
-    size_t dim, aligned_dim;
-    size_t num_points;
+    uint64_t dim, aligned_dim;
+    uint64_t num_points;
     diskann::get_bin_metadata(data_path, num_points, dim);
     aligned_dim = ROUND_UP(dim, 8);
 
@@ -187,8 +187,8 @@ void build_incremental_index(const std::string &data_path, diskann::IndexWritePa
                   << " points since the data file has only that many" << std::endl;
     }
 
-    size_t current_point_offset = points_to_skip;
-    const size_t last_point_threshold = points_to_skip + max_points_to_insert;
+    uint64_t current_point_offset = points_to_skip;
+    const uint64_t last_point_threshold = points_to_skip + max_points_to_insert;
 
     if (beginning_index_size > max_points_to_insert)
     {
@@ -243,10 +243,10 @@ void build_incremental_index(const std::string &data_path, diskann::IndexWritePa
 
         diskann::Timer timer;
 
-        for (size_t start = current_point_offset; start < last_point_threshold;
+        for (uint64_t start = current_point_offset; start < last_point_threshold;
              start += points_per_checkpoint, current_point_offset += points_per_checkpoint)
         {
-            const size_t end = std::min(start + points_per_checkpoint, last_point_threshold);
+            const uint64_t end = std::min(start + points_per_checkpoint, last_point_threshold);
             std::cout << std::endl << "Inserting from " << start << " to " << end << std::endl;
 
             auto insert_task = std::async(std::launch::async, [&]() {
@@ -277,13 +277,13 @@ void build_incremental_index(const std::string &data_path, diskann::IndexWritePa
     }
     else
     {
-        size_t last_snapshot_points_threshold = 0;
-        size_t num_checkpoints_till_snapshot = checkpoints_per_snapshot;
+        uint64_t last_snapshot_points_threshold = 0;
+        uint64_t num_checkpoints_till_snapshot = checkpoints_per_snapshot;
 
-        for (size_t start = current_point_offset; start < last_point_threshold;
+        for (uint64_t start = current_point_offset; start < last_point_threshold;
              start += points_per_checkpoint, current_point_offset += points_per_checkpoint)
         {
-            const size_t end = std::min(start + points_per_checkpoint, last_point_threshold);
+            const uint64_t end = std::min(start + points_per_checkpoint, last_point_threshold);
             std::cout << std::endl << "Inserting from " << start << " to " << end << std::endl;
 
             load_aligned_bin_part(data_path, data, start, end - start);
@@ -297,7 +297,7 @@ void build_incremental_index(const std::string &data_path, diskann::IndexWritePa
                     get_save_filename(save_path + ".inc-", points_to_skip, points_to_delete_from_beginning, end);
                 index->save(save_path_inc.c_str(), false);
                 const double elapsedSeconds = save_timer.elapsed() / 1000000.0;
-                const size_t points_saved = end - points_to_skip;
+                const uint64_t points_saved = end - points_to_skip;
 
                 std::cout << "Saved " << points_saved << " points in " << elapsedSeconds << " seconds ("
                           << points_saved / elapsedSeconds << " points/second)\n";
@@ -333,7 +333,7 @@ int main(int argc, char **argv)
     std::string data_type, dist_fn, data_path, index_path_prefix;
     uint32_t num_threads, R, L, num_start_pts;
     float alpha, start_point_norm;
-    size_t points_to_skip, max_points_to_insert, beginning_index_size, points_per_checkpoint, checkpoints_per_snapshot,
+    uint64_t points_to_skip, max_points_to_insert, beginning_index_size, points_per_checkpoint, checkpoints_per_snapshot,
         points_to_delete_from_beginning, start_deletes_after;
     bool concurrent;
