+++ extern/diskann/DiskANN/apps/utils/bin_to_tsv.cpp
@@ -5,13 +5,13 @@
 #include "utils.h"
 
 template <class T>
-void block_convert(std::ofstream &writer, std::ifstream &reader, T *read_buf, size_t npts, size_t ndims)
+void block_convert(std::ofstream &writer, std::ifstream &reader, T *read_buf, uint64_t npts, uint64_t ndims)
 {
     reader.read((char *)read_buf, npts * ndims * sizeof(float));
 
-    for (size_t i = 0; i < npts; i++)
+    for (uint64_t i = 0; i < npts; i++)
     {
-        for (size_t d = 0; d < ndims; d++)
+        for (uint64_t d = 0; d < ndims; d++)
         {
             writer << read_buf[d + i * ndims];
             if (d < ndims - 1)
@@ -41,18 +41,18 @@ int main(int argc, char **argv)
     uint32_t ndims_u32;
     reader.read((char *)&npts_u32, sizeof(uint32_t));
     reader.read((char *)&ndims_u32, sizeof(uint32_t));
-    size_t npts = npts_u32;
-    size_t ndims = ndims_u32;
+    uint64_t npts = npts_u32;
+    uint64_t ndims = ndims_u32;
     std::cout << "Dataset: #pts = " << npts << ", # dims = " << ndims << std::endl;
 
-    size_t blk_size = 131072;
-    size_t nblks = ROUND_UP(npts, blk_size) / blk_size;
+    uint64_t blk_size = 131072;
+    uint64_t nblks = ROUND_UP(npts, blk_size) / blk_size;
 
     std::ofstream writer(argv[3]);
     char *read_buf = new char[blk_size * ndims * 4];
-    for (size_t i = 0; i < nblks; i++)
+    for (uint64_t i = 0; i < nblks; i++)
     {
-        size_t cblk_size = std::min(npts - i * blk_size, blk_size);
+        uint64_t cblk_size = std::min(npts - i * blk_size, blk_size);
         if (type_string == std::string("float"))
             block_convert<float>(writer, reader, (float *)read_buf, cblk_size, ndims);
         else if (type_string == std::string("int8"))
