+++ extern/diskann/DiskANN/apps/utils/int8_to_float.cpp
@@ -13,7 +13,7 @@ int main(int argc, char **argv)
     }
 
     int8_t *input;
-    size_t npts, nd;
+    uint64_t npts, nd;
     diskann::load_bin<int8_t>(argv[1], input, npts, nd);
     float *output = new float[npts * nd];
     diskann::convert_types<int8_t, float>(input, output, npts, nd);
